pyproject.toml
[project]
name = "sentiment-analysis"
version = "3.0.0"
description = "Bitcoin news sentiment analysis tool with AI-powered insights"
authors = [
    {name = "plomitt", email = "46419727+plomitt@users.noreply.github.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "instructor (>=1.11.3,<2.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "scipy (>=1.16.2,<2.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "pydantic (>=2.5.0,<3.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "psycopg[binary,pool] (>=3.2.12,<4.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "telethon (>=1.42.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
